+++ CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file. The format 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-Nothing
+### Added
+- Add setting to auto-start Taskbar Notifier
+- Allow changing the notification location
+- Allow closing the notification with any mouse event
+- Add setting to flash the screen shortly on notifications
 
 ## [0.3.0] - 2020-12-08
 ### Changed